Extracted logging hooks into a fastify plugin
