+++ package.json
@@ -54,6 +54,7 @@
     "@tryghost/referrer-parser": "0.1.7",
     "dotenv": "16.5.0",
     "fastify": "5.4.0",
+    "fastify-plugin": "^5.0.1",
     "pino": "9.7.0",
     "ua-parser-js": "1.0.40"
   }
